clean up and test
