+++ scripts/PMPdriver_EnsoMetrics.py
@@ -16,8 +16,10 @@ import sys
 from collections import defaultdict
 from genutil import StringConstructor
 from pcmdi_metrics.driver.pmp_parser import PMPParser
-from pmpParser import ReadOptions
-from pmpParser import sort_human
+from PMPdriver_lib import ReadOptions
+from PMPdriver_lib import metrics_to_json
+from PMPdriver_lib import sort_human
+from PMPdriver_lib import tree
 from EnsoMetrics.EnsoCollectionsLib import CmipVariables, defCollection, ReferenceObservations
 from EnsoMetrics.EnsoComputeMetricsLib import ComputeCollection
 
@@ -70,13 +72,12 @@ outdir = StringConstructor(str(outdir_template(
     mip=mip, exp=exp, metricsCollection=mc_name)))
 netcdf_path = outdir(output_type='diagnostic_results')
 json_name_template = param.process_templated_argument("json_name")
-json_name = json_name_template(mip=mip, exp=exp, metricsCollection=mc_name)
+netcdf_name_template = param.process_templated_argument("netcdf_name")
 
 print('outdir:', str(outdir_template(
     output_type='%(output_type)',
     mip=mip, exp=exp, metricsCollection=mc_name))) 
 print('netcdf_path:', netcdf_path)
-print('json_name:', json_name)
 
 # Switches
 debug = param.debug
@@ -185,77 +186,6 @@ for obs in list_obs:
 print('PMPdriver: dict_obs readin end')
 
 # =================================================
-# Prepare outputing metrics to JSON file
-# -------------------------------------------------
-# Dictionary to save result 
-def tree(): return defaultdict(tree)
-
-def metrics_to_json(dict_obs, dict_metric, dict_dive, json_name, mod=None, run=None):
-    # disclaimer and reference for JSON header
-    disclaimer = open(
-        os.path.join(
-            egg_pth,
-            "disclaimer.txt")).read()
-    
-    if mc_name == 'MC1':
-        reference = "The statistics in this file are based on Bellenger, H et al. Clim Dyn (2014) 42:1999-2018. doi:10.1007/s00382-013-1783-z"
-    elif mc_name == 'ENSO_perf':
-        reference = "MC for ENSO Performance..."
-    elif mc_name == 'ENSO_tel':
-        reference = "MC for ENSO Teleconnection..."
-    elif mc_name == 'ENSO_proc':
-        reference = "MC for ENSO Process..."
-    else:
-        reference = mc_name
-    
-    enso_stat_dic = tree() # Use tree dictionary to avoid declearing everytime
-
-    # First JSON for metrics results
-    enso_stat_dic['obs'] = dict_obs
-    if mod is not None and run is not None:
-        enso_stat_dic['model'][mod][run] = dict_metric[mod][run]
-        json_name = '_'.join([json_name, mod, run])
-    else:
-        enso_stat_dic['model'] = dict_metric
-    metrics_dictionary = collections.OrderedDict()
-    metrics_dictionary["DISCLAIMER"] = disclaimer
-    metrics_dictionary["REFERENCE"] = reference
-    metrics_dictionary["RESULTS"] = enso_stat_dic
-    
-    OUT = pcmdi_metrics.io.base.Base(outdir(output_type='metrics_results'), json_name+'.json')
-    OUT.write(
-        metrics_dictionary,
-        json_structure=["type", "data", "metric", "item", "value or description"],
-        indent=4,
-        separators=(
-            ',',
-            ': '),
-        sort_keys=True)
-    
-    # Second JSON for dive down information
-    diveDown_dictionary = collections.OrderedDict()
-    diveDown_dictionary["DISCLAIMER"] = disclaimer
-    diveDown_dictionary["REFERENCE"] = reference
-    diveDown_dictionary["RESULTS"] = {}
-    if mod is not None and run is not None:
-        diveDown_dictionary["RESULTS"]["model"] = {}
-        diveDown_dictionary["RESULTS"]["model"][mod] = {}
-        diveDown_dictionary["RESULTS"]["model"][mod][run] = {}
-        diveDown_dictionary["RESULTS"]["model"][mod][run] = dict_dive[mod][run]
-    else:
-        diveDown_dictionary["RESULTS"]["model"] = dict_dive
-    
-    OUT2 = pcmdi_metrics.io.base.Base(outdir(output_type='metrics_results'), json_name+'_diveDown.json')
-    OUT2.write(
-        dict_dive,
-        json_structure=["type", "data", "metric", "item", "value or description"],
-        indent=4,
-        separators=(
-            ',',
-            ': '),
-        sort_keys=True)
-
-# =================================================
 # Loop for Models 
 # -------------------------------------------------
 # finding file and variable name in file for each observations dataset
@@ -361,12 +291,15 @@ for mod in models:
                 """
 
                 # Prepare netcdf file setup
-                netcdf_name = StringConstructor(param.netcdf_name)(model=mod, realization=run)
+                json_name = json_name_template(mip=mip, exp=exp, metricsCollection=mc_name, model=mod, realization=run)
+                netcdf_name = netcdf_name_template(mip=mip, exp=exp, metricsCollection=mc_name, model=mod, realization=run)
                 netcdf = os.path.join(netcdf_path, netcdf_name)
+
                 if debug:
                     print('file_name:', file_name)
                     print('list_files:', list_files)
                     print('netcdf_name:', netcdf_name)
+                    print('json_name:', json_name)
 
                 # Computes the metric collection
                 dict_metric[mod][run], dict_dive[mod][run] = ComputeCollection(mc_name, dictDatasets, mod, netcdf=param.nc_out,
@@ -379,7 +312,7 @@ for mod in models:
                     print(json.dumps(dict_metric, indent=4, sort_keys=True))
 
                 # OUTPUT METRICS TO JSON FILE
-                metrics_to_json(dict_obs, dict_metric, dict_dive, json_name, mod=mod, run=run)
+                metrics_to_json(dict_obs, dict_metric, dict_dive, outdir, json_name, mod=mod, run=run)
 
             except Exception as e: 
                 print('failed for ', mod, run)
@@ -392,4 +325,5 @@ print('PMPdriver: model loop end')
 # =================================================
 # OUTPUT METRICS TO JSON FILE
 # -------------------------------------------------
-#metrics_to_json(dict_obs, dict_metric, dict_dive, json_name)
+json_name = json_name_template(mip=mip, exp=exp, metricsCollection=mc_name, model='all', realization='all')
+metrics_to_json(dict_obs, dict_metric, dict_dive, outdir, json_name)
