+++ scripts/my_Param_cmip5_ENSO_perf.py
@@ -73,5 +73,5 @@ results_dir = os.path.join(
     pmprdir,
     '%(output_type)', 'enso_metric',
     '%(mip)', '%(exp)', case_id, '%(metricsCollection)')
-json_name = '_'.join(['%(mip)_%(exp)_%(metricsCollection)', case_id])
-netcdf_name = json_name + '_%(model)' + '_%(realization)'
+json_name = '_'.join(['%(mip)_%(exp)_%(metricsCollection)', case_id, '%(model)', '%(realization)'])
+netcdf_name = json_name
